SDL_PrivateJoystickForceRecentering(): fix infinite loop

+++ src/joystick/SDL_joystick.c
@@ -2492,7 +2492,7 @@ bool SDL_IsJoystickBeingAdded(void)
 
 void SDL_PrivateJoystickForceRecentering(SDL_Joystick *joystick)
 {
-    Uint8 i, j;
+    int i, j;
     Uint64 timestamp = SDL_GetTicksNS();
 
     SDL_AssertJoysticksLocked();
